pyproject.toml
[tool.poetry]
name = "langchain-project-first"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.19"
python-dotenv = "^1.0.1"
langchain-openai = "^0.3.6"
langserve = {extras = ["all"], version = "^0.3.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
